requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptocode==0.1
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
importlib-metadata==6.8.0
isort==5.12.0
Markdown==3.4.4
psycopg2-binary==2.9.7
pycparser==2.21
pycryptodomex==3.18.0
PyJWT==2.8.0
python-creditcard==0.0.1
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
zipp==3.16.2
